requirements.txt
torchvision
sagemarker
boto3
torch
numpy
pandas
scikit-learn
matplotlib
pillow
jupyterlab